refractor: job handler
