+++ public/JobHandler/components/DetailLink.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+import { EuiLink } from "@elastic/eui";
+import { ROUTES } from "../../utils/constants";
+
+export const DetailLink = (props: { index: string; writingIndex?: string }) => {
+  return <EuiLink href={`#${ROUTES.INDEX_DETAIL}/${props.writingIndex ? props.writingIndex : props.index}`}>{props.index}</EuiLink>;
+};
